+++ packages/one-app-server-bundler/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/one-app-server-bundler@1.0.3...@americanexpress/one-app-server-bundler@1.1.0) (2024-04-02)
+
+
+**Note:** Version bump only for package @americanexpress/one-app-server-bundler
+
+
+
+
+
+## [1.0.4](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/one-app-server-bundler@1.0.3...@americanexpress/one-app-server-bundler@1.0.4) (2024-03-29)
+
+
+### Bug Fixes
+
+* transpile cjs files ([#633](https://github.com/americanexpress/one-app-cli/issues/633)) ([c5e3968](https://github.com/americanexpress/one-app-cli/commit/c5e3968221aba52d95e94a0da273ef6fb4d0ec36))
+
+
+### Performance Improvements
+
+* **one-app-server-bundler:** serve-module read and write the module map once ([#623](https://github.com/americanexpress/one-app-cli/issues/623)) ([391a390](https://github.com/americanexpress/one-app-cli/commit/391a390d2c27d4343763d9a5e109b9056302ce3d))
+
+
+
+
+
+## [1.0.3](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/one-app-server-bundler@1.0.2...@americanexpress/one-app-server-bundler@1.0.3) (2024-03-26)
+
+
+### Bug Fixes
+
+* **one-app-server-bundler:** include legacy bundle on serve-module even if it does not exist ([#629](https://github.com/americanexpress/one-app-cli/issues/629)) ([f48ef4f](https://github.com/americanexpress/one-app-cli/commit/f48ef4f3f0516e8fb3dc9e4246507b10feb47ca5))
+
+
+
+
+
 ## [1.0.2](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/one-app-server-bundler@1.0.1...@americanexpress/one-app-server-bundler@1.0.2) (2024-03-15)